src/sharepoint_bulk_ingestion/requirements.txt
#ragstack-ai==0.8.0
Office365-REST-Python-Client~=2.5.5
requests~=2.31.0
textract~=1.6.5
langchain~=0.1.7
langchain-astradb~=0.0.1
azure-identity~=1.15.0
azure-keyvault-secrets~=4.8.0
boto3~=1.34.59
azure-storage-blob~=12.19.1